Fix and overhaul `exec -i word`
This changes the behavior of e.g. `exec -s -i word -o buffer tr A-z .`
(when there's no selection) to match that of `-i line` and `-i buffer`
(i.e. to replace the text used as input). It also cleans up the code
and fixes a bug (assertion failure) that could be triggered with the
following sequence of commands:

    open -t
    insert -m 'AAA BBB'
    left
    exec -s -i word -o eval sh -c 'echo show env `cat`'

Note that this is a breaking change to a documented command, but the
previous behavior was rather questionable and it seems likely no one
will notice given that no one has complained.