Simplify `u_skip_chars()`
There's only a single caller of this function and it doesn't make
use of `width` as an out-param, so it can just be passed by value
instead.

The comments were also confusing, bordering on useless and only
applicable to the previous version. A useful/descriptive comment
will be added in a later commit.