Clarify comments above line getter functions in `src/block-iter.[ch]`
The "bi should be at bol" comment above `block_iter_get_line_with_nl()`
made it ambiguous as to whether this is a hard requirement. In practice,
this function has the useful behavior of returning the *remainder* of
the line (starting from `bi`) and this aspect is being used by at least
`do_search_fwd()`.