Add `dte -Z` flag, for printing `buildvar_string[]` to `stdout`
This is a list of names for compile-time optional features. It could
previously be logged via `$DTE_LOG`, but having a simple command-line
flag to print it is much more convenient.

Note: this has been left undocumented for now, which implies it may
still receive breaking changes in the future.