Fix compiler warnings produced when building with Android NDK
Specifically, with the Clang compiler shipped in Android NDK r27c
(current LTS version).

Warning 1:

>    src/main.c:620:22: error: use of logical '||' with constant
>    operand [-Werror,-Wconstant-logical-operand]
>      620 |     if ((DEBUG >= 1) || (ASAN_ENABLED == 1) || ...) {
>          |                      ^  ~~~~~~~~~~~~~~~~~~~
>    src/main.c:620:22: note: use '|' for a bitwise operation

Warning 2 (repeated many more times in `build/gen/builtin-config.h`):

>    build/gen/builtin-config.h:3:8: error: an attribute list cannot
>    appear here
>        3 | static CONFIG_SECTION const char builtin_rc[] =
>          |        ^~~~~~~~~~~~~~

The extra command added to `.gitlab-ci.yml` is for the sake of showing
basic problems with the compiler as early and clearly as possible and
isn't specifically related to the above compiler warnings.