Allow `compile` and `tag` commands to use 3 separate message lists
The motivating use case for this is to allow the `tag` command to be
used without overwriting the previous `compile` command's messages.
This is done by having 3 different lists and allowing them to be
addressed separately with `-A`, `-B` and `-C` flags.

Usage example:

    # Push 3 messages into list -A
    compile -1s -A basic printf 'MSG1\nMSG2\nMSG3\n'

    # Push 3 messages (when in the dte repo) into list -C
    tag -C main

    # Both lists are retained separately and can now be navigated
    # using `msg [-n|-p]` with the new `[-A|-B|-C]` flags
    msg -An
    msg -Cn

    # Using the existing `msg -n` is equivalent to `msg -An`
    msg -n

    # Instead of explicitly using `tag -C main`, the new `msg-tag`
    # option can be used to make `tag` equivalent to `tag -C` by
    # default
    set msg-tag C
    tag xmalloc xstrdup
    msg -Cn

Summary of additions:

* `tag [-A|-B|-C]`
* `compile [-A|-B|-C]`
* `msg [-A|-B|-C]`
* `set msg-tag C` (sets the default message list used by `tag`,
  `exec -o tag` and `dte -t`)
* `set msg-compile B` (sets the default message list used by `compile`)
* `show msgC`
* `exec -i msgB -o msgB`
* `exec -o tagC`
* `$MSGPOS_C`
* New `msg -n` and `msg -p` key bindings for lists `-B` and `-C`

This commit also does some semi-related refactoring in `src/show.c`
and `test/dump.c`.

Note: this has been left undocumented for now, which implies it may
still receive breaking changes in the future.