Move target-specific parts of `mk/compiler.sh` to `mk/cc-target.sh`
This produces a new, generated makefile in `build/gen/cc-target.mk`.

`CC_TARGET` (and the variables derived from it) should be updated
when `CFLAGS` changes, so it makes sense to do this separately
from `build/gen/compiler.mk` (which only depends on `CC`).

This commit also fixes building with the Android NDK for API level
24 (see the added comment).