Revert most of the previous commit and clean up `mk/compiler.sh`
This moves target-specific variables back to `mk/compiler.sh`, but
now omitting any use of `CFLAGS`.

The reason the previous commit stated:

> `CC_TARGET` (and the variables derived from it) should be updated
> when `CFLAGS` changes

... was that different `CFLAGS` could, in principle, change the target
triple. However, this is never the case with GCC and is typically not
the way cross-compilation would be done with Clang either. Making the
new `build/gen/cc-target.mk` makefile depend on `build/gen/all.cflags`
was also causing redundant updates in full builds:

    $ make -B
      UPDATE  build/gen/cc-version.txt
         GEN  build/gen/platform.mk
      UPDATE  build/gen/all.cflags
         GEN  build/gen/cc-target.mk
         GEN  build/gen/compiler.mk
      UPDATE  build/gen/all.cflags
         GEN  build/gen/cc-target.mk
      UPDATE  build/gen/all.ldflags

Since there's a cleaner way to do the `-liconv` part of the detection
(as mentioned in commit 04c3da8f71e6), we dodge this (mostly imagined)
issue for now and just assume `CFLAGS` doesn't affect `CC_TARGET`.