Don't enter `main_loop()` if a `dte -c` command has already `quit`
This commit is more or less reverting 2f3e3793d (and thus un-reverting
fec217333), but done in the context of 3654a57ae (which was committed
more recently and helps avoid the edge cases mentioned in 2f3e3793d).

The patch here is somewhat different to fec217333, because some related
parts of `main()` were reorganized for "headless" mode (in 3dcea9ae7).

See also:

* The comment below `fast_exit` added to `src/main.c`
* Commit fec217333bf8e9de1a5c77a8aa851b954d486cac
* Commit 2f3e3793d607c2607d7afa1139fa10c738c9b25f
* Commit 3654a57ae113c209b44edfb67ad79c1b12ee3619
* Commit 3dcea9ae7f6aa62d86e99fe02a20981c4ba212bb