Add support for embedding and executing built-in scripts
Summary of user-visible changes:

* A new `-b` flag has been added to the `exec` and `compile` commands,
  which makes them look up (and execute, if found) the command argument
  as a built-in script (instead of in the filesystem)
* The example scripts previously found in the `contrib/` directory have
  been moved to `config/script/` and are now embedded into the editor
  (in the same way as the existing `dterc(5)` and `dte-syntax(5)` files)
* The example aliases previously found in `contrib/README.md` have moved
  to a new, embedded `extra` config and the 2 example key bindings have
  been converted to aliases
* Command-line auto-completion has been implemented for `exec -b` and
  `compile -b`
* The `show builtin` sub-command has been updated to use the correct
  filetype (and syntax highlighter) when viewing built-in scripts

This feature is implemented on top of the POSIX `fexecve(3)` function
and the Linux/FreeBSD `memfd_create(2)` function, so it only works on
those 2 platforms for now. As mentioned in the added `TODO` comment,
a more portable fallback implementation (based on POSIX `shm_open(3)`)
should be possible and will most likely be added in future commits.

Implementing it this way may seem somewhat awkward, but 2 of my main
goals are:

1. The editor itself must be self-contained (excluding man pages,
   shell completions, etc.)
2. Scripts that can be executed from the filesystem must also be
   embeddable, with no changes required

With these goals in mind, this seemed like the natural way to implement
it.

Note that everything added in this commit has been left undocumented,
which implies it may still receive breaking changes.