Add `errorfmt` `-c` flag, for clearing existing entries before appending
This can be useful as a concise way to make config files "idempotent"
(i.e. loadable multiple times, without appending an ever-growing number
of `errorfmt` entries). This was already done by the built-in configs,
using the pattern:

    errorfmt c
    errorfmt c regexp ...

This new `-c` flag makes most sense for use cases where only 1 entry
is needed and the above method would be needlessly verbose.