Make built-in `msgs.sh` script highlight whole lines in `fzf` preview
Only if the terminal does "back color erase" for the `EL` sequence.
Otherwise it should behaves as before (highlighting just the text
contents).

The main benefit of doing this is that messages for empty lines (as
often seen when running `git-changes`) will be highlighted in full,
instead of not at all.