Check for integer overflow in `calloc()` arguments, if `$CC` is pre-C23
The `xcalloc()` wrapper function already did this, but there are some
other parts of the codebase that call `calloc()` directly. This commit
moves the code previously in `xcalloc()` out to a generic helper and
uses it to check the other `calloc()` callers. It also renames some
confusingly named (but otherwise unrelated) parameters in the affected
functions.