Replace `unsigned char*` parameters and struct members with `char*`
This was a holdover from dex and made e.g. `buf[0]` produce a value with
`unsigned char` type, so that no intermediate casts were needed before
converting to wider types. It was also an obscure detail that wasn't
always obvious at the point of use and necessitated using extra compiler
flags to suppress constraint violation warnings.