Move `term_get_features()` and related code to `src/terminal/feature.c`
So that other headers (e.g. `src/terminal/color.h`) can include it
(instead of `src/terminal/terminal.h`) without producing circular
dependencies.

This commit also renames the generated `build/gen/feature.h` header
to `build/gen/build-defs.h` and updates various references to it.
Even though this is being done now to avoid ambiguity with the new
`src/terminal/feature.h` header, it makes sense in general for it
to have a globally unique name (i.e. with or without an explicit
path prefix). One specific reason for this is so that the logic in
`tools/hdrcheck.awk` is free of special cases.