Handle `line.length == 0` in `line_has_opening_brace()`
`line.data` can (in theory) be `NULL` in this case and would be passed
to `regexp_exec()`, which is marked with a `NONNULL_ARG(2)` attribute.
It would then subsequently be passed to `regexec(3)`, where a `NULL`
argument would be UB.

This isn't actually possible in any current caller, but it makes sense
to handle it, in case there are future changes like e.g. those made to
`insert_text()` and `line_contents_increases_indent()` in commit
5a7f9dae183a42.