Remove Slackware package from "installing" section of `README.md`
There seems to have been an attempt (by a bot) to update the package
to release `1.11.1`, which failed with the error:

> dte: ERR: dte.SlackBuild:28: VERSION (1.9.1) doesn't match VERSION
> in the .info file (1.11.1)

As far as I can tell this was a real flaw in the PR, which could be
fixed by bumping the `VERSION` variable in `dte.SlackBuild`. However,
the PR was closed and the package still remains at `1.9.1`. That
release is too old to mention in `README.md` at this point, so just
remove it for now.

See also:

* https://github.com/SlackBuildsOrg/slackbuilds/pull/8406
* https://github.com/SlackBuildsOrg/slackbuilds/blob/c69c628bf51de0b976a61ff42788f04fa80cdcba/development/dte/dte.SlackBuild#L28
* https://github.com/SlackBuildsOrg/slackbuilds/tree/c69c628bf51de0b976a61ff42788f04fa80cdcba/development/dte