Add `insert` command to search and command modes
This is similar to the normal mode command of the same name, but
inserts into the command-line buffer and only supports the `-m`
flag. Currently the only way to use this is via key bindings.

Note: this has been left undocumented for now, which implies it may
still receive breaking changes in the future.